refactor(core): one cross-encoder reranker module; break the rerankers import cycle
bge-reranker-v2-m3.ts and ms-marco-minilm-l6.ts were ~120 lines each and
differed only in model id and adapter name, so a fix to tokenization or
logit handling could land in one and miss the other. One factory now builds
both; the pipeline cache is keyed by model id so behaviour is unchanged.

rerankers/index.ts re-exported fit-check, which imports reranker-eval, which
imports hybrid-search, which imports rerankers/index — a runtime ESM cycle
whose only symptom would be a TDZ crash at module instantiation. Core's
index now imports checkRerankerFit from fit-check directly; the cycle is
gone (import-cycle count 2 -> 1, the remainder is type-only).
